+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Performance improvement to use less the `nextToken()` function (#397)
 * Lexer - Solving ambiguity on function keywords (#385)
+* Drop PHP 7.1 support
 
 ## [5.6.0] - 2023-01-02